+++ R/ranger.R
@@ -22,7 +22,7 @@ vetiver_ptype.ranger <- function(model, ...) {
     rlang::check_dots_used()
     dots <- list(...)
     check_ptype_data(dots)
-    ptype <- vctrs::vec_ptype(dots$ptype_data)
+    ptype <- vctrs::vec_ptype(dots$prototype_data)
     tibble::as_tibble(ptype)
 }
 
@@ -36,7 +36,7 @@ handler_startup.ranger <- function(vetiver_model) {
 #' @export
 handler_predict.ranger <- function(vetiver_model, ...) {
 
-    ptype <- vetiver_model$ptype
+    ptype <- vetiver_model$prototype
 
     function(req) {
         new_data <- req$body