+++ tests/testthat/test-pin-read-write.R
@@ -7,35 +7,38 @@ test_that("can pin a model", {
         pin_read(b, "cars1"),
         list(
             model = butcher::butcher(cars_lm),
-            ptype = vctrs::vec_slice(tibble::as_tibble(mtcars[,2:3]), 0),
+            prototype = vctrs::vec_slice(tibble::as_tibble(mtcars[,2:3]), 0),
             required_pkgs = NULL
         )
     )
 })
 
-test_that("can pin a model with no ptype", {
+test_that("can pin a model with no prototype", {
     b <- board_temp()
-    v <- vetiver_model(cars_lm, "cars_null", save_ptype = FALSE)
+    expect_snapshot_warning(
+        v <- vetiver_model(cars_lm, "cars_null", save_ptype = FALSE)
+    )
+    v <- vetiver_model(cars_lm, "cars_null", save_prototype = FALSE)
     vetiver_pin_write(b, v)
     expect_equal(
         pin_read(b, "cars_null"),
         list(
             model = butcher::butcher(cars_lm),
-            ptype = NULL,
+            prototype = NULL,
             required_pkgs = NULL
         )
     )
 })
 
-test_that("can pin a model with custom ptype", {
+test_that("can pin a model with custom prototype", {
     b <- board_temp()
-    v <- vetiver_model(cars_lm, "cars_custom", save_ptype = mtcars[3:10, 2:3])
+    v <- vetiver_model(cars_lm, "cars_custom", save_prototype = mtcars[3:10, 2:3])
     vetiver_pin_write(b, v)
     expect_equal(
         pin_read(b, "cars_custom"),
         list(
             model = butcher::butcher(cars_lm),
-            ptype = mtcars[3:10, 2:3],
+            prototype = mtcars[3:10, 2:3],
             required_pkgs = NULL
         )
     )
@@ -79,7 +82,7 @@ test_that("can read a pinned model", {
              url = meta$local$url,
              required_pkgs = v$metadata$required_pkgs)
     )
-    expect_equal(v1$ptype, v$ptype)
+    expect_equal(v1$prototype, v$prototype)
     expect_equal(v1$versioned, FALSE)
 })
 
@@ -103,6 +106,6 @@ test_that("can read a versioned model with metadata", {
              url = meta$local$url,
              required_pkgs = v$metadata$required_pkgs)
     )
-    expect_equal(v4$ptype, v$ptype)
+    expect_equal(v4$prototype, v$prototype)
     expect_equal(v4$versioned, TRUE)
 })